Reorg other files.
