+++ xcp_d/workflows/anatomical/__init__.py
@@ -1,10 +1,11 @@
 """Workflow for anatomical postprocessing."""
 
-from xcp_d.workflows.anatomical import outputs, parcellation, surface, volume
+from xcp_d.workflows.anatomical import outputs, parcellation, plotting, surface, volume
 
 __all__ = [
     "outputs",
     "parcellation",
+    "plotting",
     "surface",
     "volume",
 ]
